player db seeds

+++ db/schema.rb
@@ -28,7 +28,7 @@ ActiveRecord::Schema.define(version: 20170925195543) do
     t.string "rank", null: false
     t.string "aga_number", null: false
     t.date "membership_exp_date", null: false
-    t.decimal "rating", null: false
+    t.integer "rating", null: false
     t.string "chapter_affiliation", null: false
     t.string "state", null: false
     t.datetime "created_at", null: false
